requirements.txt
-r code/requirements.txt
Pillow>=10.0.0
requests>=2.31.0
pycryptodome>=3.20.0
